refactor: Add v2/v3 swap switch. V3 currently unimplemented.
